# src/main/resources/static/api/openapi.yml
openapi: 3.1.0
info:
  title: BOS API documentation
  description: |
    내꿈은건물주  API 문서

  version: 0.0.1-SNAPSHOT
  contact: {}
...
paths:
...
  /users/me:
    get:
      tags:
        - user
      operationId: getMe
      summary: "내 정보 조회"
      description: |-
          유효한 액세스토큰으로 조회한 유저의 프로필 정보 조회
      security:
          - BearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/GetMeResponseSpec"
        "401":
          description: 인증 실패
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/BaseProblemDetails"
    patch:
      tags:
        - user
      operationId: updateMe
      summary: "내 정보 수정"
      description: |-
        유저의 프로필 정보를 수정합니다.
        - 닉네임, 캐릭터 컴포넌트, 홈 타입, 알림 허용 여부, 마케팅 동의 여부를 수정할 수 있습니다.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserProfileRequest"
      responses:
        "200":
          $ref: "#/components/responses/GetMeResponseSpec"
        "400":
          $ref: "#/components/responses/BadRequestResponseSpec"
        "401":
          description: 인증 실패
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/BaseProblemDetails"
  /users/me/password:
    post:
      tags:
        - user
      operationId: changePassword
      summary: "비밀번호 변경"
      description: |-
        로그인한 사용자의 비밀번호를 변경합니다.

        ## 사용 방법
        - 현재 비밀번호와 새로운 비밀번호를 입력하여 변경 요청
        - 현재 비밀번호가 일치해야 변경 가능
        - 새 비밀번호는 현재 비밀번호와 달라야 함

        ## 비밀번호 요구사항
        - 8~20자 이내
        - 영문 대소문자, 숫자, 특수문자 중 3가지 이상 조합
        - 연속된 문자나 숫자 3자 이상 금지

        ## Request Body
        - currentPassword: 현재 비밀번호
        - newPassword: 새로운 비밀번호
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - currentPassword
                - newPassword
              properties:
                currentPassword:
                  type: string
                  description: 현재 비밀번호
                  example: "CurrentPassword123!"
                newPassword:
                  type: string
                  description: 새로운 비밀번호
                  example: "NewPassword123!"
                  minLength: 8
                  maxLength: 20
      responses:
        "204":
          description: |-
            비밀번호 변경이 성공적으로 완료되었습니다. 응답 바디 없이 204만 반환합니다.
        "400":
          description: |-
            비밀번호 변경 요청 실패

            **에러 케이스:**
            - `INVALID_PASSWORD`: 현재 비밀번호가 올바르지 않습니다
            - `PASSWORD_POLICY_VIOLATION`: 새 비밀번호가 요구사항을 만족하지 않습니다
            - `SAME_PASSWORD`: 새 비밀번호가 현재 비밀번호와 동일합니다
            - `INVALID_REQUEST`: 요청 형식이 올바르지 않습니다
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorCode:
                    type: string
                    example: "INVALID_PASSWORD"
                  message:
                    type: string
                    example: "현재 비밀번호가 올바르지 않습니다."
        "401":
          description: |-
            인증 실패

            **에러 케이스:**
            - `INVALID_TOKEN`: 유효하지 않은 토큰
            - `USER_NOT_FOUND`: 사용자를 찾을 수 없음
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/BaseProblemDetails"
        "500":
          $ref: "#/components/responses/InternalServerErrorResponseSpec"
...
  /transactions:
    post:
      tags:
        - transaction
      operationId: createTransaction
      summary: "거래내역 생성"
      description: |-
        새로운 거래내역을 생성합니다.

        ## 거래 유형
        - LEND: 갚을돈 (내가 빌려준 돈)
        - BORROW: 받을돈 (내가 빌린 돈)

        ## 관계 유형
        - PARENT: 부모님
        - SIBLING: 형제자매
        - FRIEND: 친구
        - LOVER: 애인
        - OTHER: 기타 (customRelationship 필드 사용)

        ## 상환 유형
        - BY_DATE: 날짜로 상환
        - BY_MONTHLY: 월별 상환
        - FLEXIBLE: 유연한 상환
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionRequest"
      responses:
        "201":
          description: 거래내역 생성 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponseSpec"
        "401":
          description: 인증 실패
        "500":
          $ref: "#/components/responses/InternalServerErrorResponseSpec"
  /transactions/{id}:
    get:
      tags:
        - transaction
      operationId: getTransactionDetail
      summary: "거래내역 조회"
      description: |-
        특정 거래내역의 상세 정보를 조회합니다.

        - 본인이 생성한 거래내역만 조회 가능합니다.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: 거래 ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 거래내역 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        "401":
          description: 인증 실패
        "404":
          description: 거래내역을 찾을 수 없음
    put:
      tags:
        - transaction
      operationId: updateTransaction
      summary: "거래내역 수정"
      description: |-
        특정 거래내역의 정보를 수정합니다.
        - 본인이 생성한 거래내역만 수정 가능합니다.
        - 상대방 이름, 캐릭터, 관계, 메모만 수정 가능합니다.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: 거래 ID
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTransactionRequest"
      responses:
        "200":
          description: 거래내역 수정 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponseSpec"
        "401":
          description: 인증 실패
        "404":
          description: 거래내역을 찾을 수 없음
    delete:
      tags:
        - transaction
      operationId: deleteTransaction
      summary: "거래내역 삭제"
      description: |-
        특정 거래내역을 삭제합니다.
        - 본인이 생성한 거래내역만 삭제 가능합니다.
        - 거래내역 삭제 시 연관된 상환 일정도 모두 삭제됩니다.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: 거래 ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: 거래내역 삭제 성공
        "401":
          description: 인증 실패
        "404":
          description: 거래내역을 찾을 수 없음
  /transactions/{id}/repayment-schedules:
    get:
      tags:
        - transaction
      operationId: getRepaymentManagement
      summary: "상환 관리 조회"
      description: |-
        특정 거래내역의 상환 관리 정보를 조회합니다.

        ## 응답 구조
        - **overdueRepayments**: 지연된 상환 목록 (예정일이 지난 미완료 상환)
        - **regularRepayments**: 정규 상환 목록 (예정, 진행중, 완료된 상환)

        ## 상환 상태별 표시 정보
        - **OVERDUE** (지연): 예정일(이미 지난 날짜), 예정금액
        - **SCHEDULED** (예정): 예정일(미래 날짜), 예정금액
        - **IN_PROGRESS** (진행중): 예정일, 예정금액
        - **COMPLETED** (완료): 실제 완료일, 실제 금액
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: 거래 ID
          example: 123
      responses:
        "200":
          description: 상환 관리 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepaymentManagementResponse"
        "401":
          description: 인증 실패
        "404":
          description: 거래내역을 찾을 수 없음
        "500":
          $ref: "#/components/responses/InternalServerErrorResponseSpec"
  /transactions/{transactionId}/schedules/{scheduleId}/repayments:
    post:
      tags:
        - transaction
      operationId: addRepayment
      summary: "상환 처리"
      description: |-
        특정 거래내역의 특정 회차에 대한 상환을 처리합니다.
        - 본인이 생성한 거래내역에만 상환 처리를 할 수 있습니다.
        - 상환 처리는 실제로 돈을 갚았을 때 등록합니다.
        - scheduleId는 상환관리 조회 API에서 획득할 수 있습니다.
      security:
        - BearerAuth: []
      parameters:
        - name: transactionId
          in: path
          description: 거래 ID
          required: true
          schema:
            type: integer
            format: int64
        - name: scheduleId
          in: path
          description: 상환 스케줄 ID
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRepaymentRequest"
      responses:
        "201":
          description: 상환 처리 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepaymentScheduleItem"
        "400":
          $ref: "#/components/responses/BadRequestResponseSpec"
        "401":
          description: 인증 실패
        "404":
          description: 거래내역을 찾을 수 없음
  /transactions/summary:
    get:
      tags:
        - transaction
      operationId: getTransactionSummary
      summary: "거래 요약 조회"
      description: |-
        사용자의 갚을돈과 받을돈에 대한 요약 정보를 조회합니다.

        ## 응답 정보
        - 갚을돈: 총 금액, 현재까지 상환 금액, 잔여 금액
        - 받을돈: 총 금액, 현재까지 받은 금액, 잔여 금액
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 거래 요약 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebtSummaryResponse"
        "401":
          description: 인증 실패
        "500":
          $ref: "#/components/responses/InternalServerErrorResponseSpec"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    BaseProblemDetails:
      type: object
      required:
        - type
        - title
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        detail:
          type: string
...
    ProfileAsset:
      type: object
      properties:
        id:
          type: string
          description: 프로필 컴포넌트의 키
          example: "FACE_TYPE_1"
        uri:
          type: string
          format: uri
          description: 프로필 컴포넌트의 이미지 URI
          example: "https://example.com/face_type_1.svg"
    Character:
      type: object
      properties:
        face:
          $ref: "#/components/schemas/ProfileAsset"
        hand:
          $ref: "#/components/schemas/ProfileAsset"
        skinColor:
          type: string
          description: 피부색
          example: #FFDBAC
        bang:
          $ref: "#/components/schemas/ProfileAsset"
        backHair:
          $ref: "#/components/schemas/ProfileAsset"
        eyes:
          $ref: "#/components/schemas/ProfileAsset"
        mouth:
          $ref: "#/components/schemas/ProfileAsset"
        home:
          $ref: "#/components/schemas/ProfileAsset"
    UpdateProfileCharacter:
      type: object
      properties:
        face:
          type: string
          example: "FACE_TYPE_1"
        hand:
          type: string
          example: "HAND_TYPE_1"
        skinColor:
          type: string
          description: 피부색
          example: "#FFDBAC"
        bang:
          type: string
          example: "BANG_TYPE_1"
        backHair:
          type: string
          example: "BACK_HAIR_TYPE_1"
        eyes:
          type: string
          example: "EYES_TYPE_1"
        mouth:
          type: string
          example: "MOUTH_TYPE_1"
        home:
          type: string
          example: "HOME_TYPE_1"
    UpdateUserProfileRequest:
      type: object
      properties:
        nickname:
          type: string
          description: 유저의 닉네임
          example: "홍길동"
        character:
          $ref: "#/components/schemas/UpdateProfileCharacter"
        isNotificationAllowed:
          type: boolean
          description: 유저가 알림을 허용하는지 여부
          example: true
        isMarketingAgreed:
          type: boolean
          description: 유저가 마케팅 수신에 동의하는지 여부
          example: true
    UserProfileResponse:
      type: object
      required:
        - id
        - nickname
        - email
        - characterComponents
        - isNotificationAllowed
        - isMarketingAgreed
      properties:
        id:
          type: integer
          format: int64
          description: 유저 아이디
          example: 15423
        email:
          $ref: "#/components/schemas/Email"
        nickname:
          type: string
          description: 유저의 닉네임
          example: "홍길동"
        character:
          $ref: "#/components/schemas/Character"
        isNotificationAllowed:
          type: boolean
          description: |-
              유저가 알림을 허용하는지 여부
          example: true
        isMarketingAgreed:
          type: boolean
          description: |-
              유저가 마케팅 수신에 동의하는지 여부
          example: true
...
    Email:
      type: string
      format: email
      description: email 주소
      example: abc@abc.com
...
    TransactionType:
      type: string
      enum:
        - LEND
        - BORROW
      description: |-
        거래 유형
        - LEND: 갚을돈 (내가 빌려준 돈)
        - BORROW: 받을돈 (내가 빌린 돈)
      example: LEND
    RelationshipType:
      type: string
      enum:
        - PARENT
        - SIBLING
        - FRIEND
        - LOVER
        - OTHER
      description: |-
        관계 유형
        - PARENT: 부모님
        - SIBLING: 형제자매
        - FRIEND: 친구
        - LOVER: 애인
        - OTHER: 기타 (customRelationship 필드 사용)
      example: FRIEND
    RepaymentType:
      type: string
      enum:
        - DIVIDED_BY_PERIOD
        - FIXED_MONTHLY
        - FLEXIBLE
      description: |-
        상환 유형
        - DIVIDED_BY_PERIOD: 기간별 분할 상환 (목표일까지 균등 분할)
        - FIXED_MONTHLY: 고정 월납 상환 (매월 고정 금액)
        - FLEXIBLE: 유연한 상환 (자유 납부)
      example: FIXED_MONTHLY
    RepaymentStatus:
      type: string
      enum:
        - SCHEDULED
        - IN_PROGRESS
        - OVERDUE
        - COMPLETED
      description: |-
        상환 상태
        - SCHEDULED: 예정됨 (납부일이 오늘 이후)
        - IN_PROGRESS: 진행중 (납부일이 2일 전~오늘)
        - OVERDUE: 연체됨 (납부일이 2일 전 이전)
        - COMPLETED: 완료됨 (실제 납부 완료)
      example: SCHEDULED
    CreateTransactionRequest:
      type: object
      required:
        - transactionType
        - counterpartName
        - counterpartCharacter
        - relationship
        - transactionDate
        - totalAmount
        - repaymentType
      properties:
        transactionType:
          $ref: "#/components/schemas/TransactionType"
        counterpartName:
          type: string
          description: 상대방 이름
          example: "홍길동"
          maxLength: 12
        counterpartCharacter:
          $ref: "#/components/schemas/Character"
        relationship:
          $ref: "#/components/schemas/RelationshipType"
        customRelationship:
          type: string
          nullable: true
          description: 기타 관계 (relationship이 OTHER일 때 사용)
          example: "동업자"
        transactionDate:
          type: string
          format: date
          description: 거래일
          example: "2025-01-15"
        totalAmount:
          type: number
          format: decimal
          description: 총 거래 금액
          example: 100000
          minimum: 10000
        completedAmount:
          type: number
          format: decimal
          nullable: true
          description: 현재까지 완료된 금액
          example: 0
          minimum: 0
        memo:
          type: string
          nullable: true
          description: 거래 메모
          example: "급하게 빌린 돈"
        repaymentType:
          $ref: "#/components/schemas/RepaymentType"
        targetDate:
          type: string
          format: date
          nullable: true
          description: 완료 예정일
          example: "2025-03-15"
        monthlyAmount:
          type: number
          format: decimal
          nullable: true
          description: 월 납부 금액
          example: 50000
          minimum: 0
        paymentDay:
          type: integer
          nullable: true
          description: 매달 납부일 (1-31)
          example: 15
          minimum: 1
          maximum: 31
        hasTargetDate:
          type: boolean
          nullable: true
          description: 완료일 설정 여부
          example: true
    UpdateTransactionRequest:
      type: object
      required:
        - counterpartName
        - counterpartCharacter
        - relationship
      properties:
        counterpartName:
          type: string
          description: 상대방 이름
          example: "홍길동"
          maxLength: 12
        counterpartCharacter:
          $ref: "#/components/schemas/Character"
        relationship:
          $ref: "#/components/schemas/RelationshipType"
        customRelationship:
          type: string
          nullable: true
          description: 기타 관계 (relationship이 OTHER일 때 사용)
          example: "동업자"
        memo:
          type: string
          nullable: true
          description: 거래 메모
          example: "급하게 빌린 돈"
    TransactionResponse:
      type: object
      required:
        - id
        - transactionType
        - counterpartName
        - counterpartCharacter
        - relationship
        - transactionDate
        - totalAmount
        - completedAmount
        - remainingAmount
        - repaymentType
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
          description: 거래 ID
          example: 123
        transactionType:
          $ref: "#/components/schemas/TransactionType"
        counterpartName:
          type: string
          description: 상대방 이름
          example: "홍길동"
        counterpartCharacter:
          $ref: "#/components/schemas/Character"
        relationship:
          $ref: "#/components/schemas/RelationshipType"
        customRelationship:
          type: string
          nullable: true
          description: 기타 관계
          example: "동업자"
        transactionDate:
          type: string
          format: date
          description: 거래일
          example: "2025-01-15"
        totalAmount:
          type: number
          format: decimal
          description: 총 거래 금액
          example: 100000
        completedAmount:
          type: number
          format: decimal
          description: 현재까지 완료된 금액
          example: 30000
        remainingAmount:
          type: number
          format: decimal
          description: 잔여 금액
          example: 70000
        memo:
          type: string
          nullable: true
          description: 거래 메모
          example: "급하게 빌린 돈"
        repaymentType:
          $ref: "#/components/schemas/RepaymentType"
        targetDate:
          type: string
          format: date
          nullable: true
          description: 완료 예정일
          example: "2025-03-15"
        monthlyAmount:
          type: number
          format: decimal
          nullable: true
          description: 월 납부 금액
          example: 50000
        paymentDay:
          type: integer
          nullable: true
          description: 매달 납부일
          example: 15
        hasTargetDate:
          type: boolean
          nullable: true
          description: 완료일 설정 여부
          example: true
        createdAt:
          type: string
          format: date-time
          description: 생성 일시
          example: "2025-01-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          description: 수정 일시
          example: "2025-01-15T10:30:00Z"
    RepaymentScheduleItem:
      type: object
      required:
        - id
        - status
        - displayDate
        - displayAmount
      properties:
        id:
          type: integer
          format: int64
          description: 상환 스케줄 ID
          example: 456
        status:
          $ref: "#/components/schemas/RepaymentStatus"
        displayDate:
          type: string
          format: date
          description: 표시할 날짜 (지연/예정/진행중은 예정일, 완료는 실제완료일)
          example: "2025-02-15"
        displayAmount:
          type: number
          format: decimal
          description: 표시할 금액 (지연/예정/진행중은 예정금액, 완료는 실제금액)
          example: 50000
    RepaymentManagementResponse:
      type: object
      required:
        - overdueRepayments
        - regularRepayments
      properties:
        overdueRepayments:
          type: array
          items:
            $ref: "#/components/schemas/RepaymentScheduleItem"
          description: 지연된 상환 목록
        regularRepayments:
          type: array
          items:
            $ref: "#/components/schemas/RepaymentScheduleItem"
          description: 정규 상환 목록 (예정, 진행중, 완료된 상환)
    CreateRepaymentRequest:
      type: object
      required:
        - repaymentDate
        - repaymentAmount
      properties:
        repaymentDate:
          type: string
          format: date
          description: 상환일
          example: "2025-01-20"
        repaymentAmount:
          type: number
          format: decimal
          description: 상환 금액
          example: 50000
          minimum: 0.01
...
    DebtSummaryResponse:
      type: object
      required:
        - debtSummary
      properties:
        debtSummary:
          $ref: "#/components/schemas/DebtSummary"
    DebtSummary:
      type: object
      required:
        - lendSummary
        - borrowSummary
      properties:
        lendSummary:
          $ref: "#/components/schemas/TransactionSummary"
          description: 갚을돈 요약 (내가 빌려준 돈)
        borrowSummary:
          $ref: "#/components/schemas/TransactionSummary"
          description: 받을돈 요약 (내가 빌린 돈)
    TransactionSummary:
      type: object
      required:
        - totalAmount
        - completedAmount
        - remainingAmount
      properties:
        totalAmount:
          type: number
          format: decimal
          description: 총 금액
          example: 1000000
        completedAmount:
          type: number
          format: decimal
          description: 현재까지 상환/받은 금액
          example: 300000
        remainingAmount:
          type: number
          format: decimal
          description: 잔여 금액
          example: 700000
...
  responses:
...
    BadRequestResponseSpec:
      description: Bad Request
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/BaseProblemDetails"
...
    InternalServerErrorResponseSpec:
      description: 내부 오류가 생겼습니다. 해당 상황 발생 시, 관리자에게 이야기해주세요.
      content:
        application/json:
          schema:
            properties:
              statusCode:
                type: number
                example: 500
              message:
                type: string
                example: Internal Server Error
...
    GetMeResponseSpec:
      description: |-
        유효한 액세스토큰으로 조회한 유저 프로필 정보 조회
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/UserProfileResponse"
...